+++ WHO_WINS_PARSER/src/main/java/com/rozhnov/who_wins_parser/controller/ParseController.java
@@ -1,11 +1,8 @@
 package com.rozhnov.who_wins_parser.controller;
 
-import com.rozhnov.who_wins_parser.entity.Match;
 import com.rozhnov.parser.DataParsing;
 import com.rozhnov.parser.info.ParsingInfo;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -13,19 +10,11 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
-import java.net.URI;
-import java.net.URISyntaxException;
-
-import static org.springframework.http.ResponseEntity.badRequest;
 
 @Controller
 @Slf4j
 @RequestMapping("/parse")
 public class ParseController {
-
-//    @Value("${spring.application.name}")
-//    private String applicationName;
-
     DataParsing dataParsing;
 
     public ParseController() {
@@ -34,17 +23,19 @@ public class ParseController {
 
     @GetMapping("/results/{count}")
     public ResponseEntity<ParsingInfo> addResults(@PathVariable int count) {
-        log.debug("Rest request to parse " + count + "results");
-        ParsingInfo parsing = null;
-        if (count != 0) {
-            parsing = dataParsing.parseResults(count);
+        log.debug("GET request to parse " + count + "results");
+
+        if (count <= 0) {
+            return ResponseEntity.badRequest().body(null);
         }
 
+        ParsingInfo parsing = dataParsing.parseResults(count);
         return ResponseEntity.ok(parsing);
     }
 
     @GetMapping("/results/{from}/{to}")
     public ResponseEntity<ParsingInfo> addResults(@PathVariable int from, @PathVariable int to) {
+        log.debug("GET request to parse results from " + from + " to " + to);
         try {
             ParsingInfo parsing = dataParsing.parseResultsOf(from, to);
             return ResponseEntity.ok(parsing);
@@ -55,25 +46,27 @@ public class ParseController {
 
     @GetMapping("/results/today")
     public ResponseEntity<ParsingInfo> addTodayResults() {
+        log.debug("GET request to parse today results");
         ParsingInfo parsing = dataParsing.parseTodayResults();
-        return new ResponseEntity<>(parsing, HttpStatus.OK);
+        return ResponseEntity.ok(parsing);
     }
 
     @GetMapping("/results/yesterday")
     public ResponseEntity<ParsingInfo> addYesterdayResults() {
+        log.debug("GET request to parse yesterday results");
         ParsingInfo parsing = dataParsing.parseYesterdayResults();
-        return new ResponseEntity<>(parsing, HttpStatus.OK);
+        return ResponseEntity.ok(parsing);
     }
 /*
     @GetMapping("/results/")
     public ResponseEntity<ParsingInfo<Match>> checkUpdates() {
         dataParsing.checkResultsUpdates();
-        return new ResponseEntity<>(null, HttpStatus.OK);
+        return ResponseEntity.ok(parsing);
     }
 
     @GetMapping("/matches/today")
     public ResponseEntity<ParsingInfo<Match>> addTodayMatches() {
         ParsingInfo<Match> parsing = dataParsing.parseTodayMatches();
-        return new ResponseEntity<>(parsing, HttpStatus.OK);
+        return ResponseEntity.ok(parsing);
     }*/
 }